content/base/test/chrome/test_bug514705.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Bug 514705"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="RunTest();">
  <script type="application/javascript" 
          src="chrome://mochikit/content/MochiKit/packed.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=421622"
     target="_blank">Mozilla Bug 421622</a>
  </body>

  <script type="application/javascript">
  <![CDATA[
    //SimpleTest.waitForExplicitFinish();
    todo(false, "Enable this test");
    
    function RunTest()
    {
      //window.open("bug514705_helper.xul", "bug514705",
      //            "chrome,width=100,height=100");
    }
  ]]>
  </script>
</window>
